target_server might return compressed response
If target_server returns compressed response then header will indicate
content-type as gzip, br, etc but when we
return then those will be uncompressed.
This creates an issue, so decode_content=False helps.

If target sends transfer-encoding as chunked, then we suppress it since we are
reading all the chunks and returning in one go.
If we want to implement chunked encoding then we'll have to do that properly.